[Imp] Mod Export is no longer limited to 2GB files. [Mod] OpenMPT: Version is now 1.22.07.01
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@3237 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Moddoc.cpp
@@ -1793,7 +1793,7 @@ void CModDoc::OnFileWaveConvert(ORDERINDEX nMinOrder, ORDERINDEX nMaxOrder, cons
 
 		// Render song (or current channel, or current sample/instrument)
 		CDoWaveConvert dwcdlg(m_SndFile, thisName, wsdlg.m_Settings, pMainFrm);
-		dwcdlg.m_dwFileLimit = static_cast<DWORD>(wsdlg.m_dwFileLimit);
+		dwcdlg.m_dwFileLimit = wsdlg.m_dwFileLimit;
 		dwcdlg.m_bGivePlugsIdleTime = wsdlg.m_bGivePlugsIdleTime;
 		dwcdlg.m_dwSongLimit = wsdlg.m_dwSongLimit;
